Decouple documentation index from man pages
In general, the HTML documentation can (and should) contain more than
just a single “Usage” chapter. For example, we might want to add a
second man page, and include the latter in the HTML documentation, too.
Or we might want to add unrelated info to the HTML documentation which
is not supposed to show up in a man page.

To accommodate for that, add one layer of indirection so the man page
will be generated from the Usage document, and at the same time, the
Usage document will be included in the HTML documentation.